requirements.txt
Flask
requests
numpy
pytest
fastapi
uvicorn
python-dotenv